AetherFrame/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
celery==5.3.6
redis==5.0.8
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.17.2
pydantic==2.9.2
pydantic-settings==2.7.1
python-multipart==0.0.9
minio==7.2.15
pytest==7.4.4
httpx==0.27.2
cryptography==43.0.1
PyYAML>=6.0
